wd2010: mask off unused bits of cylinder high
The 3.51m Kernel uses the width of the Cylinder High register to
identify the disc controller type. dhreset() in gdhd.o first resets the
controller, then writes 0xFF to the Cylinder High register. It expects
to read 0x03 (WD1010) or 0x07 (WD2010) back.

If it reads something other than 0x03 or 0x07, the kernel ends up
spinning in a loop on the bootscreen, spamming the log with the message
"Disk Ctrlr = ?".

Closes #17