Fix P3...P5 board version check (#4)
3B1 Rev P3, P4 and P5 motherboards loop back the PIE+ (parity error
interrupt enable) status into the most significant bit of the GSR. The
kernel and diag disks use this to detect the hardware rev.

We emulate the P5.1 PAL already, but need to emulate this otherwise
'revlev' will be set incorrectly for P3..P5 hardware, preventing the use
of the P5.1 PAL.

P6, by the way, is a P5.1 motherboard with a 2400 Baud modem and the P5.1
PAL integrated onto the motherboard.